requirements.txt
backoff
mock
requests
singer_python
